+++ changelog.md
@@ -2,8 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-# [0.0.69] - In progress
+# [0.0.70] - In progress
+*
 
+# [0.0.69] - 2024-05-06
+* add toc generator for html and refactoring, Builder3
+* add static markdown generation, Builder3
+* add markdown output path, Builder3
+  
 # [0.0.68] - 2024-04-29
 * add content parser to Builder3
 * start ui library for Builder3